Refactor MetadataProvider
Extract interface and two implementations
