+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/ivyresolve/ComponentSelectionRulesProcessor.java
@@ -91,7 +91,7 @@ public class ComponentSelectionRulesProcessor {
             return Collections.emptyList();
         }
 
-        if (!metadataProvider.resolve()) {
+        if (!metadataProvider.isUsable()) {
             return null;
         }
